Move ruby files in Library folder
