+++ library/base_decorator.rb
