+++ library/capitalize_decorator.rb
@@ -1,4 +1,4 @@
-require_relative 'basedecorator'
+require_relative 'base_decorator'
 
 class CapitalizeDecorator < BaseDecorator
   def correct_name
