+++ library/nameable.rb
