+++ library/trimmer_decorator.rb
@@ -1,4 +1,4 @@
-require_relative 'basedecorator'
+require_relative 'base_decorator'
 class TrimmerDecorator < BaseDecorator
   def correct_name
     name = @nameable.correct_name
